requirements.txt
streamlit
langchain-community
langchain
pytube
python-dotenv
streamlit==1.35.0
langchain==0.1.20
langchain-groq==0.0.6
sentence-transformers==2.7.0
groq==0.6.0
tqdm==4.66.4
numpy==1.26.4
scikit-learn==1.4.2
transformers==4.41.1
torch==2.3.0
python-dotenv==1.0.1
